Lectura de bytes

+++ src/ssdfs/aux/auxiliary_fx.c
@@ -6,13 +6,13 @@
 // NOTE: El plano 0 es el primer plano
 // NOTE: El bloque 0 es el primer plano
 // ...
-int calc_offset(int plane, int block, int page, int cell, int bytes) {
-    int offset = 0;
+long int calc_offset(int plane, int block, int page, int cell, int bytes) {
+    long int offset = 0;
 
-    int cell_offset = BYTES_PER_CELL;
-    int page_offset = CELLS_PER_PAGE * cell_offset;
-    int block_offset = PAGES_PER_BLOCK * page_offset;
-    int plane_offset = BLOCKS_PER_PLANE * block_offset;
+    long int cell_offset = BYTES_PER_CELL;
+    long int page_offset = CELLS_PER_PAGE * cell_offset;
+    long int block_offset = PAGES_PER_BLOCK * page_offset;
+    long int plane_offset = BLOCKS_PER_PLANE * block_offset;
 
     offset += bytes;
     offset += cell * cell_offset;
